Fixed bug where data wasn't resetting between menus. Had to move all events into a common events class.
