+++ src/main/java/toffydungeons/toffydungeons/ToffyDungeons.java
@@ -4,6 +4,7 @@ import org.bukkit.Bukkit;
 import org.bukkit.plugin.java.JavaPlugin;
 import toffydungeons.toffydungeons.GUIs.DungeonCreationMenu;
 import toffydungeons.toffydungeons.GUIs.DungeonSelectionMenu;
+import toffydungeons.toffydungeons.GUIs.InventoryEvents;
 import toffydungeons.toffydungeons.commands.TDungeonCommand;
 
 public final class ToffyDungeons extends JavaPlugin {
@@ -18,9 +19,7 @@ public final class ToffyDungeons extends JavaPlugin {
         } else {
             System.out.println("[Toffy Dungeons]: Toffy Dungeons version " + this.getServer().getVersion() +" loaded up! ");
             this.getCommand("TDungeon").setExecutor(new TDungeonCommand());
-
-            getServer().getPluginManager().registerEvents(new DungeonSelectionMenu(), this);
-            getServer().getPluginManager().registerEvents(new DungeonCreationMenu(), this);
+            getServer().getPluginManager().registerEvents(new InventoryEvents(), this);
         }
     }
